Extract factory class and extract library load.
